refactor: couponFixture 통합
